ml-service/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
gunicorn>=21.2.0
